Added middleware for route protection

+++ app/api/products/[id]/route.ts
@@ -1,19 +1,19 @@
+import { requireAdmin } from "@/lib/api-utils";
 import { db } from "@/lib/firebase";
 import { deleteDoc, doc, updateDoc } from "firebase/firestore";
 import { NextResponse } from "next/server";
 
 export async function PUT(
   req: Request,
- { params }: { params: Promise<{ id: string }> }
+  { params }: { params: Promise<{ id: string }> },
 ) {
-
-  const {id} = await params;
-  const authHeader = req.headers.get("authorization");
+   const user = await requireAdmin();
   
-      if (!authHeader) {
-        return NextResponse.json({ error: "Unauthorized" }, { status: 401 });
-      }   
-
+      if(!user){
+        return NextResponse.json({error:"Unauthorized"},{status:401})
+      }    
+  
+  const { id } = await params;
   const data = await req.json();
 
   await updateDoc(doc(db, "products", id), data);
@@ -21,19 +21,17 @@ export async function PUT(
   return Response.json({ success: true });
 }
 
-
 export async function DELETE(
   req: Request,
-  { params }: { params: Promise<{ id: string }> }
-
+  { params }: { params: Promise<{ id: string }> },
 ) {
-
-  const {id} = await params;
-  const authHeader = req.headers.get("authorization");
+  const user = await requireAdmin();
   
-      if (!authHeader) {
-        return NextResponse.json({ error: "Unauthorized" }, { status: 401 });
-      }   
+  if(!user){
+    return NextResponse.json({error:"Unauthorized"},{status:401})
+  }    
+  
+  const { id } = await params;
 
   await deleteDoc(doc(db, "products", id));
